services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tothemoontokens-api"
version = "0.1.0"
description = "Hackathon artifact API for paid agent work, settlement, and delivery evidence."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.0,<3.0",
  "httpx>=0.27,<1.0",
  "websockets>=14.0,<16.0",
  "structlog>=24.1,<26.0",
  "prometheus-client>=0.20,<1.0",
  "cryptography>=47.0,<48.0",
  "google-genai>=1.73,<2.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "argon2-cffi>=23.1,<24.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "email-validator>=2.2,<3.0",
  "asyncpg>=0.29,<1.0",
  "psycopg[binary]>=3.1,<4.0",
  "redis>=5.0,<7.0",
  "opentelemetry-api>=1.27,<2.0",
  "opentelemetry-sdk>=1.27,<2.0",
  "opentelemetry-exporter-otlp>=1.27,<2.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0,<1.0",
  "sentry-sdk[fastapi]>=2.0,<3.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-cov>=5.0,<7.0",
  "ruff>=0.6,<1.0",
  "mypy>=1.11,<2.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tothemoon_api = ["data/*.json"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q --strict-markers"

[tool.coverage.run]
source = ["tothemoon_api"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 70

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
  "SIM",
  "RUF"
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]

[tool.mypy]
python_version = "3.12"
packages = ["tothemoon_api"]
mypy_path = "src"
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
check_untyped_defs = true
no_implicit_optional = true
disallow_untyped_decorators = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
ignore_errors = true
